refactor of tinywl and code addition for keybindings
